requirements.txt
Django==2.0rc1
djangorestframework==3.7.3
psycopg2==2.7.3.2
pytz==2017.3
